Autenticação da documentação API

+++ routes/api.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Http\Controllers\Api\V1\Admin\Auth\LoginController;
+use App\Http\Controllers\Api\V1\Admin\Auth\RegisterController;
 use App\Http\Controllers\Api\V1\Admin\ProjectsApiController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -19,14 +21,22 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::group(['prefix' => 'v1', 'namespace' => 'Api\V1\Admin'], function () {
+Route::prefix('auth')->group(function() {
+    Route::post('login', [LoginController::class, 'login']);
+    Route::post('logout', [LoginController::class, 'logout'])->middleware('auth:sanctum');
+    Route::post('register', [RegisterController::class, 'register']);
+});
 
+
+Route::group(['prefix' => 'v1', 'namespace' => 'Api\V1\Admin'], function () {
+    Route::post('login', [ProjectsApiController::class, 'login']);
     // Projects
-    Route::get('index', [ProjectsApiController::class, 'index']);
-    Route::get('show/{id}', [ProjectsApiController::class, 'show']);
-    Route::post('store', [ProjectsApiController::class, 'store']);
-    Route::put('update/{id}', [ProjectsApiController::class, 'update']);
-    Route::delete('destroy/{id}', [ProjectsApiController::class, 'destroy']);
+    Route::get('index', [ProjectsApiController::class, 'index'])->middleware('auth:sanctum');
+    Route::get('show/{id}', [ProjectsApiController::class, 'show'])->middleware('auth:sanctum');
+    Route::post('store', [ProjectsApiController::class, 'store'])->middleware('auth:sanctum');
+    Route::put('update/{id}', [ProjectsApiController::class, 'update'])->middleware('auth:sanctum');
+    Route::delete('destroy/{id}', [ProjectsApiController::class, 'destroy'])->middleware('auth:sanctum');
     
 });
+
